+++ lib/ab_admin/views/admin_navigation_helpers.rb
@@ -20,20 +20,19 @@ module AbAdmin
         if adapter && adapter.klass == resource_class
           sort_link(adapter, attribute, options)
         else
-          attribute.is_a?(Symbol) ? ha(attribute) : attribute
+          options[:title] || (attribute.is_a?(Symbol) ? ha(attribute) : attribute)
         end
       end
 
-      def sort_link(search, attribute, *args)
-        name = attribute.is_a?(Symbol) ? ha(attribute) : attribute
-        return name unless search
-
-        options = args.first.is_a?(Hash) ? args.shift.dup : {}
+      def sort_link(adapter, attribute, options={})
+        name = options[:title] || (attribute.is_a?(Symbol) ? ha(attribute) : attribute)
+        return name unless adapter
         search_params = (params[:q] || {}).to_h.with_indifferent_access
         attr_name = (options.delete(:column) || attribute).to_s
         default_order = options.delete :default_order
 
-        if existing_sort = search.sorts.detect { |s| s.name == attr_name }
+        existing_sort = adapter.sorts.detect { |s| s.name == attr_name }
+        if existing_sort
           prev_attr, prev_dir = existing_sort.name, existing_sort.dir
           current_dir = prev_attr == attr_name ? prev_dir : nil
         else