+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 * backwards incompatible changes
+  * `deprecated_utils.js` removed without replacement
   * `Settings.load_config` is deprecated, use `Settings.data` instead
   * `call_method_or_proc_on` don't accept method name as String, use Symbol instead
   * `call_method_or_proc_on` use `public_send` instead of `sent` for method call